some info about NetworkModule.kt

+++ app/src/main/java/com/example/moviesproject/domain/model/MovieDetails.kt
@@ -27,7 +27,7 @@ data class MovieDetails(
 
     val releaseDate: String,
 
-    val voteAverage: Double,
+    val voteAverage: Int,
 
     val tagline: String,
